src/main/java/io/github/tjheslin1/gateway/infrastructure/domain/eventstore/BankingEventServerBuilder.java
package io.github.tjheslin1.gateway.infrastructure.domain.eventstore;

import io.github.tjheslin1.gateway.settings.Settings;
import org.eclipse.jetty.servlet.ServletContextHandler;
import org.eclipse.jetty.servlet.ServletHolder;

public class BankingEventServerBuilder {

    private ServletContextHandler context;
    private Settings settings;

    public BankingEventServerBuilder(ServletContextHandler context, Settings settings) {
        this.context = context;
        this.context.setContextPath("/");

        this.settings = settings;
    }

    public BankingEventServerBuilder withServlet(ServletHolder servlet, String path) {
        context.addServlet(servlet, path);
        return this;
    }

    public BankingEventServer build() {
        return new BankingEventServer(context, settings.serverPort());
    }
}
